refactor: remove WeatherSettings component and update admin settings page to streamline layout and settings management + fixed bug in slide panel component
